SIMDscan

Documentation for SIMDscan.

This package provides code for doing a scan using SIMD instructions. Scans are also known as prefix operations. The Base Julia function accumulate performs the same operation.

Given a sequence $x_1, x_2, ... , x_n$, and an associative operator, $\oplus$, the the scan is

\[x_1, x_1 \oplus x_2, x_3 \oplus x_2 \oplus x_1, ... , x_n \oplus x_{n-1} \oplus \cdots \oplus x_1\]

For parallelization, it is essential that $\oplus$ be associative. The function scan_simd!(⊕, x) computes the scan of x in place.

Warnings

x must be indexed from 1 to length(x).

⊕ must be associative for scan_simd!

Multivariate Operations

There is also a method for scanning $⊕: \mathbbm{R}ᴷ→\mathbbm{R}ᴷ$. In this case, ⊕ should accept two K tuples as arguments and a K-tuple. x should be a NTuple{K,AbstractVector} where each element of the tuple is the sequence of values.

The multivariate scan can be used to simulate an AR(1) model.

T = 10
ϵ = randn(T)
# simple recursive version for comparison
y = similar(ϵ)
y[1] = ϵ[1]
α = 0.5
for t = 2:T
  y[t] = α*y[t-1] + ϵ[t]
end

# to make ⊕ associative, augment ϵ[t] with a second element that keeps track of the appropriate power of α
⊕(y,e) = (e[1] + α*y[1]*e[2], α*y[2]*e[2])
id = (0.0,1.0/α) # a left identity; ⊕(id,x) = x ∀ x
yscansimd,at = scan_simd!(⊕, (copy(ϵ), ones(T)), identity=id)
[y yscansimd]

10×2 Matrix{Float64}:
  1.9937     1.9937
  0.889665   0.889665
 -0.686971  -0.686971
 -0.798751  -0.798751
  0.148711   0.148711
  1.26027    1.26027
 -0.528677  -0.528677
  0.739108   0.739108
 -0.125447  -0.125447
 -1.17033   -1.17033

Acknowledgements

The following sources were helpful while developing this package:

(Slotin, 2023) describes an SIMD implementation for prefix sum with C code

(Nash *et al.*, 2021) has example code for a threaded scan (prefix sum) in Julia,

SIMDscan.scan_serial!

SIMDscan.scan_serial!

SIMDscan.scan_simd!

SIMDscan.scan_simd!

Functions

SIMDscan.scan_serial! — Method

scan_serial!(f::F,x::AbstractVector) where {F <: Function}

Replaces the vector x with the scan of x using f. That is,

\[\begin{bmatrix} x[1] \\ f(x[2],x[1]) \\ f(x[3],f(x[2],x[1])) \\ ⋮ \end{bmatrix} # Examples\]

jl-doctest julia> scan_serial!(+,[1,2,3,4]) 4-element Vector{Int64}: 1 3 6 10 ```

SIMDscan.scan_serial! — Method

scan_serial!(f::F,x::NTuple{K, AbstractVector{T}}) where {F <: Function, K, T}

In place scan for a function that takes two K tuples as arguments. Replaces the tuple of vectors x with the scan.

Examples

julia> scan_serial!((x,y)->(x[1]+y[1], x[2]*y[2]),([1,2,3,4],[1,2,3,4]))
([1, 3, 6, 10], [1, 2, 6, 24])

SIMDscan.scan_simd! — Method

scan_simd!(f::F, x::AbstractVector{T}, v::Val{N}=Val(8);identity::T=zero(T)) where {F, T, N}

In place scan for an associative function.

identity should be a left identity under f. That is, f(identity, y) = y for all y.

T, must be a type that can be loaded onto registers, i.e. one of SIMD.VecTypes.

f must be associative. Otherwise, this will give incorrect results.

Val(N) specifies the SIMD vector width used. The default of 8 should give good performance on CPUs with AVX512 for which 8 Float64s fill the 512 bits available.

Examples

julia> scan_simd!(+,[1,2,3,4])
4-element Vector{Int64}:
  1
  3
  6
 10

SIMDscan.scan_simd! — Method

scan_simd!(f::F, x::NTuple{K, AbstractVector{T}}, v::Val{N}=Val(8); identity::NTuple{K,T}=ntuple(i->zero(T),Val(K))) where {F, K, T, N}

In place scan for an associative function that takes two K tuples as arguments. Replaces the tuple of vectors x with the scan.

identity should be a left identity under f. That is, f(identity, y) = y for all y.

T, must be a type that can be loaded onto registers, i.e. one of SIMD.VecTypes.

f must be associative. Otherwise, this will give incorrect results.

Val(N) specifies the SIMD vector width used. The default of 8 should give good performance on CPUs with AVX512 for which 8 Float64s fill the 512 bits available.

Examples

julia> scan_simd!((x,y)->(x[1]+y[1], x[2]*y[2]),([1,2,3,4],[1,2,3,4]))
([1, 3, 6, 10], [1, 2, 6, 24])